requirements.txt
Flask
requests
beautifulsoup4
ImageHash
Pillow
sqlite3
numpy